### 2021-06-16 #3379 @whedon
[PRE REVIEW]: COVID-19 Lung Segmentation
**Submitting author:** @RiccardoBiondi (<a href="http://orcid.org/0000-0002-8682-3782">Riccardo Biondi</a>)
**Repository:** <a href="https://github.com/RiccardoBiondi/segmentation" target ="_blank">https://github.com/RiccardoBiondi/segmentation</a>
**Version:** 2.0.0
