### 2021-06-16 #3379 @danielskatz
👋 @RiccardoBiondi - thanks for your submission. Due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS.  You should hear back in a week or so.  I'll also note it's surprising to me that it's not packaged for installation via PyPI.
