### 2021-06-25 #3379 @kyleniemeyer
Hi @RiccardoBiondi, the editorial board has reviewed your submission and found it is within our scope for review. We'll now try to find an editor to review it, but most submissions are holding in a waitlist right now due to low editor availability.
