### 2021-06-26 #3379 @acolum
Hi @jmschrei,

Thank you for inviting me to review this paper. Unfortunately, I don't have
enough bandwidth at this time.

Best Regards,
Alyssa

On Thu, Jun 24, 2021 at 10:12 PM Jacob Schreiber ***@***.***>
wrote:

> Hi @acolum <https://github.com/acolum> and @stensy
> <https://github.com/stensy>, would either of you be able to serve as
> reviewers for this?
>
