### 2021-07-02 #3447 @whedon
[REVIEW]: COVID-19 Lung Segmentation
**Submitting author:** @RiccardoBiondi (<a href="http://orcid.org/0000-0002-8682-3782">Riccardo Biondi</a>)
**Repository:** <a href="https://github.com/RiccardoBiondi/segmentation" target ="_blank">https://github.com/RiccardoBiondi/segmentation</a>
**Version:** 2.0.0

    - I was also able to install `requirements.txt` by running `pip install segmentation/`
    - Would be nice to have a link to the installation instructions for `pytest` and `hypothesis`; it took me some time to make sure I was installing the right ones
    - I ended up having to reinstall `SimpleITK` after installing `pytest` and `hypothesis` for some reason
    - Yes, using the testing module

    - The syntax of how to run the tools is provided, but not the command for a working example; it would be good to have an example of what commands to run, along with paths to the actual images to run on (e.g. in `images/`)
