### 2021-07-22 #3447 @amtseng
#### Extended feedback
**Installation**
Although the instructions say to install the dependencies by running `pip install -r` followed by `python setup.py`, I actually found it much easier to install the package directly by doing `pip install segmentation/` (i.e. treating the repository as a package). I would highly recommend switching to this style of installation in the instructions.

It would be nice to have a link to the installation instructions for `pytest` and `hypothesis`. For people are not familiar with these packages, it is not immediately clear which version to install. It took me two tries before I installed the correct one.

For some strange reason, I ended up having to reinstall SimpleITK again after installing `pytest` and `hypothesis`.

**Example usage**
The built-in tests are really nice to have, and it's also great that the syntax of how to run the commands is listed. However, an actual _example_ using real images would be better. That is, instead of saying `./lung_extraction.sh path/to/TRAIN/ path/to/LUNG/` using dummy paths like `path/to/TRAIN`, please include paths to actual images (perhaps under the `images/` directory), along with the expected output. This way, the user can actually see the input and output images and verify it works as expected (instead of just running tests, which can be hard to see what's going on).
