+++ docs/source/script.rst
@@ -78,15 +78,15 @@ displayed :
 
 .. code-block:: bash
 
-  I'm Loading...
+  I m Loading...
   Loaded 20 files from /path/to/input/folder/
   *****Starting clustering*****
   Number of subsamples--> 100
   Total images --> 4000
   Centroid initialization technique-->KMEANS_RANDOM_CENTERS
-  I'm clustering...
+  I m clustering...
   100%|█████████████████████████████████████████████████████████████████████████████████████| 100/100 [00:14<00:00,  2.86s/it]
-  I'm saving...
+  I m saving...
   [DONE]
 
 All the images will be divided into N subsamples, and a K-means clustering is
