### 2021-08-31 5dc14a4
[REVISION] Add tesint minimum versions. Modified installation.
