+++ docs/source/examples/modules.rst
@@ -1,0 +1,7 @@
+Examples
+========
+
+.. toctree::
+  :maxdepth: 1
+
+  ./pipeline_workflow.ipynb
