+++ docs/source/index.rst
@@ -87,7 +87,9 @@ as '.nrrd' and run the bash or PowerShell script.
   modules
   snakemake
   script
+  examples/modules
   references
+
 
 
 Indices and tables
