+++ .github/workflows/docs.yml
@@ -34,6 +34,11 @@ jobs:
     - name: Install CTLungSeg
       run: |
         python setup.py install
+      
+    - name: Install Pandoc
+      run: |
+        sudo apt install pandoc
+  
     - name: Install docs dependencies
       run: |
         python -m pip install -r docs/requirements.txt
