+++ docs/source/conf.py
@@ -34,6 +34,7 @@ extensions = [
               'sphinx.ext.napoleon',
               'sphinx.ext.viewcode',
               'nbsphinx',
+              'sphinx.ext.mathjax',
               'IPython.sphinxext.ipython_console_highlighting']
 
 # Add any paths that contain templates here, relative to this directory.
@@ -44,7 +45,7 @@ templates_path = []
 #
 # This is also used if you do content translation via gettext catalogs.
 # Usually you set "language" from the command line for these cases.
-language = 'python'
+language = "python"
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
