### 2021-09-01 680aa02
[DOCS] Updated doc version. Fix github actions badges hyperlinks
