+++ docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2020, RiccardoBiondi, Nico Curti'
 author = 'RiccardoBiondi, Nico Curti'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.0'
+release = '2.0.0'
 master_doc = 'index'
 # -- General configuration ---------------------------------------------------
 
