### 2021-09-01 1c94485
[ENH] Add readthedocs configuration file
