+++ .readthedocs.yml
@@ -1,0 +1,7 @@
+version: 2
+formats: all
+python:
+    version: 3
+    install:
+      - requirements: docs/requirements.txt
+    system_packages: true
