### 2021-09-01 #3447 @RiccardoBiondi
Dear all,
First, thank you, @amtseng, for your precious advices.
For what concern the installation, We have modified the instruction as suggested and add the minimum required versions for pytest and hypothesis.

For what concern the example, We have add in the online documentation a jupyter notebook which explains step by step the pipeline showing the intermediate results. That because to include medical images directly in the repository will increase the size of the package.

Riccardo
