### 2021-09-16 #3447 @pps121
Below are my thoughts and suggestions:

- should include more examples of how to use the software thinking from cross domain horizons (to solve by real-world problems)
- software documentation should be more details, including OS variants.
- Should clearly mention, how this packages are better than other baselines / commonly-used packages to decide state of the field.

Thank you.
