### 2021-09-28 868a26c
[REVISION] Improved documentation. Add real data examples
