+++ paper/paper.bib
@@ -74,3 +74,12 @@
   url = {https://doi.org/10.1016/j.cmi.2020.08.003},
   language = {eng}
 }
+
+@misc{yan2020covid19,
+      title={COVID-19 Chest CT Image Segmentation -- A Deep Convolutional Neural Network Solution},
+      author={Qingsen Yan and Bo Wang and Dong Gong and Chuan Luo and Wei Zhao and Jianhu Shen and Qinfeng Shi and Shuo Jin and Liang Zhang and Zheng You},
+      year={2020},
+      eprint={2004.10987},
+      archivePrefix={arXiv},
+      primaryClass={eess.IV}
+}
