+++ docs/source/script.rst
@@ -9,14 +9,23 @@ divided into two different steps :
 - lung extraction
 - labeling
 
+The following examples will use the data previously downloaded from the public dataset.
+
 Lung Extraction
 ---------------
 
 This script allows to run the lung segmentation on the whole set of patients,
 that is the preliminary step of the GGO identification. Its implemented both for
 bash and PowerShell.
-To perform the segmentation simply organize all the CT scans in the same folder
-(input folder), and create an empty folder in which the results will be saved.
+
+In this example we will segment *coronacases_002* and *coronacases_005* patients.
+To perform the segmentation, you have to create two folers:
+
+  - input: contains only the scan to segment
+  - lung: contains the lung extraction results.
+
+
+
 Ensure that in the input folder there are only the files corresponding to the scan
 to segment. The supported input format are all the ones supported by SimpleITK_.
 you can provide as input also DICOM series, simply arrange them into one folder
@@ -26,13 +35,20 @@ Now you can simply run the following command from bash :
 
 .. code-block:: bash
 
-  lung_extraction.sh /path/to/input/folder/ /path/to/output/folder/
+  mkdir ./Examples/INPUT
+  mkdir ./Examples/LUNG
+  mv ./Examples/COVID-19-CT/coronacases_002.nii.gz ./Examples/COVID-19-CT/coronacases_005.nii.gz ./Examples/INPUT
+  lung_extraction.sh ./Examples/INPUT ./Examples/LUNG
 
 or its equivalent for powershell
 
 .. code-block:: powershell
 
-  lung_extraction.ps1 /path/to/input/folder/ /path/to/output/folder/
+  New-Item -Path "Examples" -Name "INPUT" -ItemType "directory"
+  New-Item -Path "Examples" -Name "LUNG" -ItemType "directory"
+  Move-Item -Path "Examples\COVID-19-CT\coronacases_002.nii.gz" -Destination "Examples\INPUT"
+  Move-Item -Path "Examples\COVID-19-CT\coronacases_005.nii.gz" -Destination "Examples\INPUT"
+  lung_extraction.ps1 .\Examples\INPUT .\Examples\LUNG
 
 For lung extraction, a pre-trained UNet model was used. The model and the
 code used to apply it belong to this_ repository. For more details, please
@@ -41,21 +57,23 @@ refers here_.
 Labeling
 --------
 
-Once you have isolated the lung, you can run the actual segmentation. As for
-lung extraction, simply arrange all the results of the previous script into an
-input folder; as before creating an empty folder in which the resulting labels
-will be saved in '.nrrd' format.
-Simply run the bash script:
+Once you have isolated the lung, you can run the actual segmentation.
+In this case you have to create an other empty folder, wich will contains all the results.
+As input the script requires the results of the lung extraction(in this case stored in LUNG folder).
+
+Run from bash:
 
 .. code-block:: bash
 
-  labeling.sh /path/to/input/folder/ /path/to/output/folder/
+  mkdir ./Examples/OUTPUT
+  labeling.sh ./Examples/LUNG ./Examples/OUTPUT
 
 or its equivalent for powershell
 
   .. code-block:: powershell
 
-    labeling.ps1 /path/to/input/folder/ /path/to/output/folder/
+    New-Item -Path "Examples" -Name "OUTPUT" -ItemType "directory"
+    labeling.ps1 .\Examples\LUNG .\Examples\OUTPUT
 
 This will run the segmentation by using the already estimated centroids. If you
 want to use another set of centroids, simply provide as third arguments the path
@@ -71,7 +89,7 @@ training set. Now simply the training script:
 
 .. code-block:: bash
 
[…]
