+++ docs/source/examples/examples.rst
@@ -5,3 +5,4 @@ Examples
   :maxdepth: 1
 
   ./pipeline_workflow.ipynb
+  ./body_segmentation.ipynb
