### 2021-09-28 #3447 @RiccardoBiondi
Dear @pps121 @amtseng, thank you for the precious advice.

For what concern the examples, we have added a body segmentation example.

We have added instructions on how to download and organize CT scans from a public dataset. We have provided examples using real data and removing all the dummy paths.

For the OS variants, We have added instructions for data downloading and organization and for software usage, both for bash and PowerShell.

In the end, we have included in the paper a comparison between our approach and a supervised one.
