### 2021-09-28 #3447 @jmschrei
In L7 can you change ".. a novel, non-supervised..." to "a novel, unsupervised"?

In L23 can you change "It is acute, highly contagious, viral infection..." to "It is >>an<< acute, highly contagious, viral infection..."?

In L27 can you change "In this scenario" to "In the diagnostic setting," or otherwise clarify what "this" means?
