### 2021-09-29 #3447 @RiccardoBiondi
Thank you @jmschrei.

We have made the required modifications to the paper.
A tagged release is now available. We have deposited the software on Zenodo (doi: 10.5281/zenodo.5535568 url: https://doi.org/10.5281/zenodo.5535568).

Let us know if everything is ok.
