### 2021-09-29 #3447 @danielskatz
@RiccardoBiondi - I've suggested some paper edits in https://github.com/RiccardoBiondi/segmentation/pull/2

I also don't understand "more specific and performing segmentation methods" well enough to suggest a change.  Can you explain what this means?
