### 2021-09-29 #3447 @RiccardoBiondi
Hi @danielskatz

I have uploaded the suggested edits.

I have also modified the unclear sentence, "we surmise that our pipeline can be successfully used as starting point for more specific and performing segmentation methods."  to "we surmise that our pipeline can be successfully used as a first segmentation method to be used as training for other, more specific methods".
